Split character builder workspace controls
