+++ frontend/src/components/tools/character-builder/_components/character-builder-build-look-types.ts
@@ -1,0 +1,15 @@
+import type { CharacterBuilderTraits } from '@/types/character-builder';
+
+export type TraitChoiceKey = keyof Pick<
+  CharacterBuilderTraits,
+  | 'genderPresentation'
+  | 'ageRange'
+  | 'skinTone'
+  | 'faceCues'
+  | 'hairColor'
+  | 'hairLength'
+  | 'hairstyle'
+  | 'eyeColor'
+  | 'bodyBuild'
+  | 'outfitStyle'
+>;
